+++ src/features/recipe/hooks/useRecipes.ts
@@ -7,6 +7,8 @@ import { useToast } from "@/components/ui/use-toast";
 import { createRecipe, deleteRecipe, getRecipes } from "@/features/recipe/apis";
 import { RecipePreview } from "@/features/recipe/models/recipe";
 
+import { ApiException } from "@/libs/exceptions";
+
 const RECIPE_QUERY_KEY = ["recipes"] as const;
 
 export default function useRecipes() {
@@ -27,27 +29,13 @@ export default function useRecipes() {
       router.push(`/recipe/${recipe._id}`);
       return queryClient.invalidateQueries({ queryKey: RECIPE_QUERY_KEY });
     },
-    onError: (error: Error) => {
-      const { message } = error;
-      console.error(message);
+    onError: ({ message, description }: ApiException) => {
+      toast({
+        title: message,
+        description,
+      });
 
-      if (message.includes("Invalid URL")) {
-        toast({
-          title: "올바르지 않은 URL입니다.",
-          description:
-            "자막이 사용 가능한 YouTube 요리 영상의 URL을 입력해주세요.",
-        });
-      } else if (message.includes("No recipe found")) {
-        toast({
-          title: "레시피를 찾을 수 없습니다.",
-          description: "요리와 관련된 영상의 URL을 입력해주세요.",
-        });
-      } else {
-        toast({
-          title: "레시피 생성에 실패했습니다.",
-          description: "다시 시도해주세요.",
-        });
-      }
+      console.error(message, description);
     },
   });
