+++ src/features/recipe/models/recipe.ts
@@ -33,3 +33,8 @@ export interface Ingredient {
 export interface Step {
   description: string;
 }
+
+export interface Script {
+  offset: number;
+  text: string;
+}
